Implement key prefixing
This change makes it possible to use nodes as prefixes by including
a suffix key within the Node struct and appending it appropriately
during the encoding/decoding process.

Also changes the prior behavior of padding each encoded node to an 8
byte boundary. We lose some performance by not doing this, but gain
greater space efficiency. In the future, adding the padding bytes
dynamically during the decode might recover most of the lost speed,
and would be preferable over keeping them serialized.

Bumps version to 0.3.0 because this is a breaking change to the binary
format.